Split the others too
