+++ test/extension-tests/single-root-UI/test/debugger-test-kits.ts
@@ -1,8 +1,6 @@
 import {DefaultEnvironment, expect, getFirstSystemKit} from '@test/util';
-// import sinon = require('sinon');
 import * as vscode from 'vscode';
 import CMakeTools from '@cmt/cmake-tools';
-
 
 suite('[Debug/Launch interface using Kits and Variants]', async () => {
   let testEnv: DefaultEnvironment;
@@ -29,7 +27,7 @@ suite('[Debug/Launch interface using Kits and Variants]', async () => {
   teardown(async function(this: Mocha.Context) {
     this.timeout(30000);
 
-    await vscode.workspace.getConfiguration('cmake', vscode.workspace.workspaceFolders![0].uri).update('useCMakePresets', 'always');
+    await vscode.workspace.getConfiguration('cmake', vscode.workspace.workspaceFolders![0].uri).update('useCMakePresets', 'auto');
 
     testEnv.teardown();
   });
@@ -40,40 +38,5 @@ suite('[Debug/Launch interface using Kits and Variants]', async () => {
     await vscode.commands.executeCommand('cmake.selectLaunchTarget', undefined, executablesTargets[0]);
 
     await vscode.commands.executeCommand('cmake.debugTarget');
-    // sinon.assert.calledWith(testEnv.vs_debug_start_debugging);
-  }).timeout(60000);
-});
-
-suite('[Debug/Launch interface using Presets]', async () => {
-  let testEnv: DefaultEnvironment;
-
-  setup(async function(this: Mocha.Context) {
-    this.timeout(100000);
-
-    const build_loc = 'build';
-    const exe_res = 'output.txt';
-
-    testEnv = new DefaultEnvironment('test/extension-tests/single-root-UI/project-folder', build_loc, exe_res);
-    testEnv.projectFolder.buildDirectory.clear();
-
-    await vscode.commands.executeCommand('cmake.setConfigurePreset', 'LinuxUser1');
-    await vscode.commands.executeCommand('cmake.setBuildPreset', '__defaultBuildPreset__');
-    await vscode.commands.executeCommand('cmake.setTestPreset', '__defaultTestPreset__');
-
-    expect(await vscode.commands.executeCommand('cmake.build')).to.be.eq(0);
-  });
-
-  teardown(async function(this: Mocha.Context) {
-    this.timeout(30000);
-    testEnv.teardown();
-  });
-
-  test('Test call of debugger', async () => {
-    const executablesTargets = await vscode.commands.executeCommand('cmake.executableTargets') as string[];
-    expect(executablesTargets.length).to.be.not.eq(0);
-    await vscode.commands.executeCommand('cmake.selectLaunchTarget', undefined, executablesTargets[0]);
-
-    await vscode.commands.executeCommand('cmake.debugTarget');
-    // sinon.assert.calledWith(testEnv.vs_debug_start_debugging);
   }).timeout(60000);
 });
